Move Firefox Accounts to separate utils file Update import paths for dependent files
